+++ app/src/models/posts.lua
@@ -94,7 +94,7 @@ function Posts:get_listing(filters)
 
 	local query = [[
 		a.id, a.title, a.url, a.body, a.is_self, a.over_18, a.locked, a.sub_id, a.thumbnail,
-			a.stickied, a.comments_locked,
+			a.stickied, a.comments_locked, a.is_question, a.accepted_comment_id,
 			a.created_at AS age, a.created_at,
 			c.user_name AS author,
 			s.name AS subreddit,